Enable non-root users other than the default jenkins user
By using fixuid. However, most of the work was made in order to wrap the
s6-overlay's init in a suid capable binary so it still runs as root.

We also drop support for gosu in favor of setpriv, which is now more
adopted and even comes preinstalled in newer distributions.