Allow image to run as root for Jenkinsfile agents (#63)
So far it has been prepared to run as non-root only. The only real use
case for running as root is for running as a Jenkinsfile agent. Running
the jenkins-agent itself as root is not a good idea, it's not tested and
not supported.

Some things had to be taken into account:

- HOME and USER variable can no longer be hardcoded
  - when running as non-root, fixuid properly handles the HOME variable
  - USER variable will not be available when running the image as a
    Jenkinsfile agent
- fixuid should not run, otherwise it chowns the whole /home/jenkins to
  root (which I don't see as necessary)
- /init_as_root is no longer necessary, and especially we should not
  drop the privileges as part of the entrypoint.