Stop publishing to Docker Hub
I prefer to concentrate on publishing to GitHub Container Registry, so
that I don't need to maintain two accounts and also because this way all
download counts are in one place.

As part of this change, I also removed the Docker image from Docker Hub,
so that when users try to download it again, it will fail and therefore
notice that the image is no longer available there. Otherwise, they
would keep using the old image without noticing that it's no longer
updated.