Add python3 and pipx
The python3 package was present already before it was upgraded to Ubuntu
24.04, so this will fix this regression.

It is very useful to have some version of Python 3 available upfront for
running unknown scripts and workloads.

For example, a lot of Node.js dependencies depends on Python 3 to build
native modules.

Unfortunately, adding the python3 package back to the image increases
the image size by 40MB.

This also adds pipx to the images, so that Python-based tools can be
installed globally with no headaches. The pipx package itself does not
add any meaningful size to the image.